+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 #v1.0
 * Chance for an item depends on how many times you stole it instead of every items chance being affected.
 * Replaced experience with times stolen.
+* Changed equation for determining 'Your chance'.
 
-##Known Bugs
+####Bugs
 * Some pickpocket items aren't visible in '/pickpocket items' once you've received it.
 
 #v0.9.9